LayoutTests/fast/mediastream/getUserMedia-grant-persistency.html

<!DOCTYPE HTML>
<html>
    <head>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <div id="navigateToSamePage"></div>
        <script>
promise_test((test) => {
    if (window.testRunner)
        testRunner.setUserMediaPermission(true);
    return navigator.mediaDevices.getUserMedia({audio:false, video:true}).then((stream) => {
        if (window.testRunner)
            testRunner.setUserMediaPermission(false);
        return navigator.mediaDevices.getUserMedia({audio:false, video:true});
    }).then(() => {
        return navigator.mediaDevices.getUserMedia({audio:true, video:false}).then(assert_unreached, (e) => {
            assert_equals(e.name, "NotAllowedError");
        });
    }).then(() => {
        if (window.testRunner)
            testRunner.setUserMediaPermission(true);
        return navigator.mediaDevices.getUserMedia({audio:true, video:false});
    }).then(() => {
        if (window.testRunner)
            testRunner.setUserMediaPermission(false);
        return navigator.mediaDevices.getUserMedia({audio:true, video:true});
    });
}, "Testing same page getUserMedia grant persistency");

promise_test((test) => {
    if (window.testRunner)
        testRunner.setUserMediaPermission(true);
    return navigator.mediaDevices.getUserMedia({audio:true, video:true}).then((stream) => {
        if (window.testRunner)
            testRunner.setUserMediaPermission(false);
        return navigator.mediaDevices.getUserMedia({audio:true, video:true});
    }).then(() => {
        if (window.internals)
            internals.setPageMuted("capturedevices");
        return navigator.mediaDevices.getUserMedia({audio:true, video:true});
    }).then(() => {
        if (!window.internals)
            return;
        assert_true(internals.pageMediaState().includes("HasActiveVideoCaptureDevice"), "Check active video");
        assert_true(internals.pageMediaState().includes("HasActiveAudioCaptureDevice"), "Check active audio");
    });
}, "Testing same page getUserMedia grant persistency in case page capture is in muted state");

promise_test((test) => {
    if (window.testRunner)
        testRunner.setUserMediaPermission(true);
    return navigator.mediaDevices.getUserMedia({audio:false, video:true}).then((stream) => {
        if (window.testRunner)
            testRunner.setUserMediaPermission(false);
        return navigator.mediaDevices.getUserMedia({audio:false, video:true});
    }).then(() => {
        window.location = "#navigateToSamePage";
    }).then(() => {
        return navigator.mediaDevices.getUserMedia({audio:false, video:true});
    });
}, "Testing same page getUserMedia grant persistency after in page navigation");

promise_test((test) => {
    var isPageVisible = true;
    if (window.testRunner)
        testRunner.setUserMediaPermission(true);
    return navigator.mediaDevices.getUserMedia({audio:false, video:true}).then((stream) => {
        if (window.testRunner)
            testRunner.setUserMediaPermission(false);
        isPageVisible = false;
        if (window.internals)
            internals.setPageVisibility(false);
        setTimeout(() => {
            isPageVisible = true;
            if (window.internals)
                internals.setPageVisibility(true);
        }, 100);
        return navigator.mediaDevices.getUserMedia({audio:false, video:true});
    }).then(() => {
        assert_true(isPageVisible, "Resolving getUserMedia promise should wait for page to be visible");
    });
}, "Testing same page getUserMedia grant persistency with visibility");

        </script>
    </body>
</html>
